skills/skill-factory/requirements.txt
fastmcp>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
